Linux Makefile changes: Still not working
Problems with vtk #define functions.  See config.qmake to see the misbehaving defines.  Wihtout the defines, the vtk static New() is not working. There is a workaround by having the vtk-wrapping built and using the VTK_AUTOLOAD_PATH environment variable.  But then the QVTKWidget does not not work and segfault.

Everything is fine on MS visual studio.